added tests for the custom extension that enabled default reuse identifier, removed the stub AppListingTests
