--- AppListingTests/AppListingTests.swift
@@ -1,23 +1,0 @@
-//
-//  AppListingTests.swift
-//  AppListingTests
-//
-//  Created by MC on 1/3/2018.
-//  Copyright © 2018 MC. All rights reserved.
-//
-
-import XCTest
-@testable import AppListing
-
-class AppListingTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-}
